feat(sell): add SellPriceEntry model, sell_prices state, f64 coins

+++ backend/src/websocket/actions/director_action_scheduler.rs
@@ -24,7 +24,7 @@ pub struct DirectorActionParams {
     pub player_id: Option<String>,
     pub life: Option<i32>,     // 玩家生命值
     pub strength: Option<i32>, // 玩家体力值
-    pub coins: Option<i32>,    // 玩家货币
+    pub coins: Option<f64>,    // 玩家货币
     pub target_place: Option<String>,
     pub action_type: Option<String>, // rope/unrope
     pub rest_enabled: Option<bool>,  // 夜晚结算时静养是否生效
